Last-Use Early Register Release (#220)
This commit introduces the last-use early register release.

In this mechanism, physical registers are freed once the last instruction reading the register
has committed if the redefining instruction becomes non-speculative. A last-use table is employed
to track the last consumer.

This technique was first introduced in:
"Hardware schemes for early register release," in ICPP, IEEE, 2002.

This ALGO will do the register early release if the following holds:
(1) the redefine-instruction of the producer needs to be precommitted
(2) the last-use instruction of the producer needs to be committed