Update Confidence at FTs Pushed to DFE (#328)
#### Summary

This refactor changes how the confidence mechansim interacts with the
decoupled frontend. Instead of calling `conf->update()` for every op as
the decoupled frontend builds an FT, conf->update() is now called when
an FT is completed. The FT is passed to the update, and the confidence
mechanism iterates through the ops within the FT internally.

---------

Co-authored-by: Surim Oh <72247751+5surim@users.noreply.github.com>