Atomic register early release (#356)
In this mechanism, a physical register is freed once all of its consumer instructions
have executed when there are not conditional branches or exception-causing instructions
between the renaming instruction and the redefining instruction. A 3-bit counter is
used to track the number of pending consumers.

This technique was first introduced in:
    "ATR: Out-of-Order Register Release Exploiting Atomic Regions"
    Y. Zhao, S. Oh, M. Xu, and H. Litz
    (MICRO) International Symposium on Microarchitecture, 2025